+++ tests/integration/internal/tests/api/metrics/sandbox_list_metrics_test.go
@@ -1,4 +1,4 @@
-package api
+package metrics
 
 import (
 	"net/http"
