+++ tests/integration/internal/tests/api/metrics/team_metrics_test.go
@@ -1,4 +1,4 @@
-package api
+package metrics
 
 import (
 	"net/http"
