+++ tests/integration/internal/tests/api/sandboxes/sandbox_auto_pause_test.go
@@ -1,4 +1,4 @@
-package api
+package sandboxes
 
 import (
 	"net/http"
