+++ tests/integration/internal/tests/api/sandboxes/sandbox_internet_test.go
@@ -1,4 +1,4 @@
-package api
+package sandboxes
 
 import (
 	"net/http"
