+++ tests/integration/internal/tests/api/sandboxes/sandbox_list_test.go
@@ -1,4 +1,4 @@
-package api
+package sandboxes
 
 import (
 	"fmt"
