+++ tests/integration/internal/tests/api/sandboxes/sandbox_test.go
@@ -1,4 +1,4 @@
-package api
+package sandboxes
 
 import (
 	"context"
